+++ FluentPlayableApi/Packages/com.studentutu.fluentplayableapi/Tests/Editor/TestPlayableBehaviour.cs
@@ -1,0 +1,10 @@
+#nullable enable
+
+using UnityEngine.Playables;
+
+namespace FluentPlayableApi.Tests
+{
+    internal sealed class TestPlayableBehaviour : PlayableBehaviour
+    {
+    }
+}