+++ FluentPlayableApi/Packages/com.studentutu.fluentplayableapi/CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.4]
+
+### Added
+
+- Added generic fluent output creation and source declaration for any `IPlayableOutput`.
+- Added edit-mode coverage for `ScriptPlayableOutput` and `AudioPlayableOutput`.
+
+### Changed
+
+- Renamed the public C# namespace and assembly casing to `FluentPlayableApi`.
+
 ## [1.0.3] - 2026-05-05
 
 ### Changed